1227第一次提交

+++ app/src/main/java/com/example/vlayoutapplication/bean/HomeBannerBean.java
@@ -382,7 +382,7 @@ public class HomeBannerBean {
             private String pic_url;
             private int sort_order;
             private int is_show;
-            private int floor_price;
+            private float floor_price;
             private String app_list_pic_url;
             private int is_new;
             private String new_pic_url;
@@ -444,11 +444,11 @@ public class HomeBannerBean {
                 this.is_show = is_show;
             }
 
-            public int getFloor_price() {
+            public float getFloor_price() {
                 return floor_price;
             }
 
-            public void setFloor_price(int floor_price) {
+            public void setFloor_price(float floor_price) {
                 this.floor_price = floor_price;
             }
 
@@ -516,7 +516,7 @@ public class HomeBannerBean {
             private String item_pic_url;
             private String subtitle;
             private int topic_category_id;
-            private int price_info;
+            private float price_info;
             private String read_count;
             private String scene_pic_url;
             private int topic_template_id;
@@ -580,11 +580,11 @@ public class HomeBannerBean {
                 this.topic_category_id = topic_category_id;
             }
 
-            public int getPrice_info() {
+            public float getPrice_info() {
                 return price_info;
             }
 
-            public void setPrice_info(int price_info) {
+            public void setPrice_info(float price_info) {
                 this.price_info = price_info;
             }
 
@@ -683,7 +683,7 @@ public class HomeBannerBean {
                 private int id;
                 private String name;
                 private String list_pic_url;
-                private int retail_price;
+                private float retail_price;
 
                 public int getId() {
                     return id;
@@ -709,11 +709,11 @@ public class HomeBannerBean {
                     this.list_pic_url = list_pic_url;
                 }
 
-                public int getRetail_price() {
+                public float getRetail_price() {
                     return retail_price;
                 }
 
-                public void setRetail_price(int retail_price) {
+                public void setRetail_price(float retail_price) {
                     this.retail_price = retail_price;
                 }
             }
